+++ lib/nodejieba.h
@@ -8,6 +8,7 @@ extern NAN_METHOD(cut);
 extern NAN_METHOD(cutAll);
 extern NAN_METHOD(cutHMM);
 extern NAN_METHOD(cutForSearch);
+extern NAN_METHOD(cutSmall);
 extern NAN_METHOD(tag);
 extern NAN_METHOD(extract);
 extern NAN_METHOD(insertWord);