+++ test/demo.js
@@ -35,5 +35,5 @@ nodejieba.insertWord("男默女泪");
 result = nodejieba.cut("男默女泪");
 console.log(result);
 
-result = nodejieba.cut("南京市长江大桥", "MP", 3);
+result = nodejieba.cutSmall("南京市长江大桥", 3);
 console.log(result);